Re-hook most of the form logic
